private route problem solve

+++ src/component/MyToys/MyToys.jsx
@@ -4,6 +4,7 @@ import MyToysTable from "./MyToysTable";
 import { AuthContext } from "../AuthProvider/AuthProvider";
 import { ToastContainer, toast } from "react-toastify";
 import "react-toastify/dist/ReactToastify.css";
+import swal from "sweetalert";
 
 const MyToys = () => {
   useDynamicTitle("Toy Zone | My Toys");
@@ -25,18 +26,35 @@ const MyToys = () => {
   }, [user]);
 
   const handelDeleteNew = (_id) => {
-    notify("Please Wait");
-    fetch(`https://jisan-repo-production.up.railway.app/users/${_id}`, {
-      method: "DELETE",
-    })
-      .then((res) => res.json())
-      .then((data) => {
-        console.log(data);
-        if (data.deletedCount > 0) {
-          notify("Delete Success");
-          setMtToy((prevMyToys) => prevMyToys.filter((toy) => toy._id !== _id));
-        }
-      });
+    swal({
+      title: "Are you sure?",
+      text: "Do You Wan't to delete this item?",
+      icon: "warning",
+      buttons: true,
+      dangerMode: true,
+    }).then((willDelete) => {
+      if (willDelete) {
+        fetch(`https://jisan-repo-production.up.railway.app/users/${_id}`, {
+          method: "DELETE",
+        })
+          .then((res) => res.json())
+          .then((data) => {
+            console.log(data);
+            if (data.deletedCount > 0) {
+              notify("Delete Success");
+              setMtToy((prevMyToys) =>
+                prevMyToys.filter((toy) => toy._id !== _id)
+              );
+            }
+          });
+
+        swal("Delete Success!", {
+          icon: "success",
+        });
+      } else {
+        swal("Delete Cancel");
+      }
+    });
   };
 
   return (
@@ -58,6 +76,7 @@ const MyToys = () => {
         {/* head */}
         <thead>
           <tr>
+            <th>Toy Image</th>
             <th>Toy Name</th>
             <th>Seller</th>
             <th>Gmail</th>
